[MUI] Moves 'withRoot' HOC to utils folder
